move weather to own component
